requirements.txt
flask>=2.0.0
jinja2>=3.0.0
requests>=2.28.0
python-dotenv>=0.20.0